[PRM-1482] Updated Penalties Connector with new parameter (#208)
* [PRM-1482] Updated Penalties Connector with new parameter

* [PRM-1482] Removed featureswitching from ConnectorISpec

+++ app/models/v3/lpp/LPPDetails.scala
@@ -116,7 +116,7 @@ object LPPDetails extends JsonUtils {
 
 case class LPPDetailsMetadata(
                                mainTransaction: Option[MainTransactionEnum.Value],
-                               outstandingAmount: Option[Int]
+                               outstandingAmount: Option[BigDecimal]
                              )
 
 object LPPDetailsMetadata {
